read/write DB

+++ sandbox/config_selection.py
@@ -1,14 +1,14 @@
-import os
 import shutil
 import re
 import math
 import multiprocessing
 
 from env_config import *
-from spearmint.schema import Experiment as Exp
+from spearmint.schema import JobInfo
 
 from spearmint import main as bo
 import openstack_api
+from experiment import VirtualMachine, Experiment
 
 
 def clean_path(path):
@@ -34,16 +34,24 @@ def copy_tree(old_path, new_path):
     return True
 
 
-def replace_exp(old_path, new_path, file_name, exp_name, job_id):
+def replace_exp(old_path, new_path, file_name, experiment):
     read_file = open(os.path.join(old_path, file_name), "r")
     write_file = open(os.path.join(new_path, file_name), "w")
 
-    key_name = "template"
-    key_id = "job_id"
-    for line in read_file.readlines():
-        new_line = re.sub(key_name, exp_name, line)
-        new_line = re.sub(key_id, job_id, new_line)
-        write_file.write(new_line)
+    if file_name == "config.pb":
+        for line in read_file.readlines():
+            new_line = re.sub("template", experiment.name, line)
+            write_file.write(new_line)
+
+    elif file_name == "experiment_config.py":
+        for line in read_file.readlines():
+            new_line = re.sub("template", experiment.job_id, line)
+            new_line = re.sub("user_id", experiment.user_id, new_line)
+            new_line = re.sub("data_size", str(experiment.data_group), new_line)
+            write_file.write(new_line)
+
+    read_file.close()
+    write_file.close()
 
 
 def rename_exp(path, exp_name):
@@ -52,17 +60,27 @@ def rename_exp(path, exp_name):
     os.rename(template_name, dst_name)
 
 
-def create_experiment(job_id, exp_name, exp_path):
-    copy_tree(TEMPLATE_PATH, exp_path)
-    rename_exp(exp_path, exp_name)
-    replace_exp(TEMPLATE_PATH, exp_path, "config.pb", exp_name, job_id)
-    replace_exp(TEMPLATE_PATH, exp_path, "experiment_config.py", exp_name, job_id)
+def check_or_create_experiment(experiment):
 
-    # write the new experiment into DB
-    try:
-        new_experiment = Exp(name=exp_name, count=0)
-    except Exception, e:
-        print e
+    exp_path = os.path.join(EXP_PATH, experiment.name)
+    print exp_path
+    if not os.path.exists(exp_path):
+        copy_tree(TEMPLATE_PATH, exp_path)
+
+        rename_exp(exp_path, experiment.name)
+        replace_exp(TEMPLATE_PATH, exp_path, "config.pb", experiment)
+        replace_exp(TEMPLATE_PATH, exp_path, "experiment_config.py", experiment)
+
+        # write the new experiment into DB
+        try:
+            print experiment.job_id, experiment.user_id, experiment.cpu_percentage, experiment.io_percentage
+            new_experiment = JobInfo(name=experiment.job_id,
+                                     user_id=experiment.user_id,
+                                     cpu_percentage=experiment.cpu_percentage,
+                                     io_percentage=experiment.io_percentage)
+            print new_experiment
+        except Exception, e:
+            print e
 
 
 def start_bo(exp_name):
@@ -75,11 +93,8 @@ def start_bo(exp_name):
     print process_list
     if command_ not in process_list:
         p = multiprocessing.Process(name="BO_process", target=bo.main, args=(config_path,))
-        # print "Starting BO..."
+        print "Starting BO..."
         p.start()
-        # print p.pid
-        # print p.name
-        # bo.main(config_path)
 
 
 def get_vm_name(io_weight, cpu_weight, vcpus, ram, disk):
@@ -88,7 +103,7 @@ def get_vm_name(io_weight, cpu_weight, vcpus, ram, disk):
     min_distance = -1
     vm_name = ''
     for flavor in flavors:
-        if flavor['name'] not in flavor_space:
+        if flavor['name'] not in VirtualMachine.flavor_space:
             continue
         if flavor['vcpus'] < vcpus or flavor['ram'] / 1024 < ram or flavor['disk'] < disk:
             continue
@@ -105,13 +120,13 @@ def get_vm_name(io_weight, cpu_weight, vcpus, ram, disk):
     return vm_name
 
 
-def get_best_config(exp_name, io_weight, cpu_weight, is_to_optimize=True):
+def get_best_config(experiment, is_to_optimize=True):
     if is_to_optimize:
         file_name = "experiment.txt"
     else:
         file_name = "best_job_and_result.txt"
 
-    file_path = os.path.join(EXP_PATH, exp_name, file_name)
+    file_path = os.path.join(EXP_PATH, experiment.name, file_name)
 
     while not os.path.exists(file_path):
         continue
@@ -120,7 +135,7 @@ def get_best_config(exp_name, io_weight, cpu_weight, is_to_optimize=True):
     data = file_object.read()
 
     if is_to_optimize:
-        vm, vcpus, ram, disk, cluster_size, exp = data.split(' ')
+        vm_name, vcpus, ram, disk, cluster_size, exp = data.split(' ')
     else:
         data = data.split("Parameters: \n")
         print data
@@ -138,23 +153,28 @@ def get_best_config(exp_name, io_weight, cpu_weight, is_to_optimize=True):
             data += line
 
         vcpus, ram, disk, cluster_size = data.split(", ")
-        vm = get_vm_name(io_weight, cpu_weight, int(vcpus), int(ram), int(disk))
-    return vm, int(cluster_size), int(vcpus), int(ram), int(disk) * 10
+        vm_name = get_vm_name(experiment.io_weight, experiment.cpu_weight, int(vcpus), int(ram), int(disk))
+
+    experiment.set_best_configuration(int(cluster_size), vm_name, int(vcpus), int(ram), int(disk))
+    return experiment
 
 
-def select_configuration(user_id, job_id, data_size, timestamp, io_weight, cpu_weight, is_to_optimize=True):
+def select_configuration(experiment, is_to_optimize=True):
 
-    data_group = "0"
-    for split in data_split:
-        if float(data_size) - split < 0:
-            break
-        data_group = str(split)
+    check_or_create_experiment(experiment)
+    start_bo(experiment.name)
+    get_best_config(experiment, is_to_optimize)
 
-    exp_name = user_id + "-" + job_id + "-" + data_group + "-" + timestamp
-    exp_path = os.path.join(EXP_PATH, exp_name)
-    if not os.path.exists(exp_path):
-        create_experiment(job_id, exp_name, exp_path)
-    start_bo(exp_name)
-    vm, cluster_size, vcpus, ram, disk = get_best_config(exp_name, io_weight, cpu_weight, is_to_optimize)
+    return experiment.best_configuration
 
-    return exp_name, vm, cluster_size, vcpus, ram, disk
+
+# job_id = "terasort"
+# data_size = 0
+# user_id = "class_test1"
+#
+# print user_id, job_id
+# experiment = Experiment(job_id, user_id, data_size)
+# print experiment.job_id, experiment.user_id
+# experiment.set_io_cpu_percentage()
+#
+# check_or_create_experiment(experiment)
